+++ app/scripts/main.js
@@ -59,6 +59,7 @@ import Workflow from './components/workflows/workflow';
 import Executions from './components/executions';
 import ExecutionsOverview from './components/executions/overview';
 import ExecutionStatus from './components/executions/execution-status';
+import ExecutionLogs from './components/executions/execution-logs';
 
 import Rules from './components/rules';
 import RulesOverview from './components/rules/overview';
@@ -132,6 +133,10 @@ render((
         <Route path='executions' component={Executions}>
           <IndexRoute component={ExecutionsOverview} />
           <Route path='execution/:executionArn' component={ExecutionStatus} />
+        </Route>
+        <Route path='executions' component={Executions}>
+          <IndexRoute component={ExecutionsOverview} />
+          <Route path='execution/:executionName/logs' component={ExecutionLogs} />
         </Route>
         <Route path='rules' component={Rules}>
           <IndexRoute component={RulesOverview} />